refactor(infra): Move router HTTP modules into infra/http
