+++ auth/src/features/api/v1/auth/reset_password.rs
@@ -8,12 +8,14 @@ use token::services::{PasswordResetTokenService, SessionService};
 
 use crate::{
     domain::users,
-    infra::error::{AppError, AppErrorKind},
-    routers::{
-        extractor::{Json, LoginAccess},
-        response::app_error_to_response,
-        serializer::{Response, ResponseCode},
-        utils::cookie::CookieJarExt,
+    infra::{
+        error::{AppError, AppErrorKind},
+        http::{
+            extractor::{Json, LoginAccess},
+            response::app_error_to_response,
+            serializer::{Response, ResponseCode},
+            utils::cookie::CookieJarExt,
+        },
     },
     state::AppState,
 };
