+++ auth/src/features/api/v1/common/ping.rs
@@ -1,4 +1,4 @@
-use crate::routers::serializer::{Response, ResponseCode};
+use crate::infra::http::serializer::{Response, ResponseCode};
 
 pub async fn controller() -> Response {
     Response::new(ResponseCode::OK.into(), "pong".into(), None)
