+++ auth/src/features/api/v1/users/delete.rs
@@ -6,12 +6,14 @@ use token::services::SessionService;
 
 use crate::{
     domain::{permission, role, users},
-    infra::error::{AppError, AppErrorKind},
-    routers::{
-        extractor::{Json, LoginAccess, OperatorAccess},
-        response::app_error_to_response,
-        serializer::{Response, ResponseCode},
-        utils::cookie::CookieJarExt,
+    infra::{
+        error::{AppError, AppErrorKind},
+        http::{
+            extractor::{Json, LoginAccess, OperatorAccess},
+            response::app_error_to_response,
+            serializer::{Response, ResponseCode},
+            utils::cookie::CookieJarExt,
+        },
     },
     state::AppState,
 };
