+++ auth/src/features/api/v1/users/setting.rs
@@ -7,11 +7,11 @@ use crate::{
     infra::{
         database::ModelError,
         error::{AppError, AppErrorKind},
-    },
-    routers::{
-        extractor::{LoginAccess, OperatorAccess},
-        response::app_error_to_response,
-        serializer::{Response, ResponseCode},
+        http::{
+            extractor::{LoginAccess, OperatorAccess},
+            response::app_error_to_response,
+            serializer::{Response, ResponseCode},
+        },
     },
     state::AppState,
 };
