+++ auth/src/infra/http/cors.rs
