+++ auth/src/infra/http/serializer/common.rs
