+++ auth/src/infra/http/serializer/error.rs
