+++ auth/src/infra/http/serializer/mod.rs
